+++ python/solders/__init__.py
@@ -1,4 +1,6 @@
 from solders.solders import (  # type: ignore
+    account_decoder,
+    commitment_config,
     errors,
     hash,
     instruction,
@@ -8,12 +10,16 @@ from solders.solders import (  # type: ignore
     presigner,
     pubkey,
     signature,
+    rpc,
     transaction,
+    transaction_status,
 )
 from . import sysvar
 from . import system_program
 
 __all__ = [
+    "account_decoder",
+    "commitment_config",
     "errors",
     "hash",
     "instruction",
@@ -22,8 +28,10 @@ __all__ = [
     "null_signer",
     "presigner",
     "pubkey",
+    "rpc",
     "signature",
     "transaction",
+    "transaction_status",
     "sysvar",
     "system_program",
 ]