+++ src/keypair.rs
@@ -1,11 +1,12 @@
 use pyo3::{prelude::*, types::PyBytes};
+use serde::{Deserialize, Serialize};
 use solana_sdk::signer::{
     keypair::{
         keypair_from_seed, keypair_from_seed_phrase_and_passphrase, Keypair as KeypairOriginal,
     },
     Signer as SignerTrait,
 };
-use solders_macros::{common_magic_methods, pyhash, richcmp_signer};
+use solders_macros::{common_methods, pyhash, richcmp_signer};
 
 use crate::{
     handle_py_value_err, impl_display, impl_signer_hash, pubkey::Pubkey, signature::Signature,
@@ -13,8 +14,28 @@ use crate::{
     ToSignerOriginal,
 };
 
+mod keypair_serde {
+    use serde::{self, Deserialize, Deserializer, Serializer};
+    use solana_sdk::signer::keypair::Keypair as KeypairOriginal;
+
+    pub fn serialize<S>(kp: &KeypairOriginal, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serializer.serialize_bytes(&kp.to_bytes())
+    }
+
+    pub fn deserialize<'de, D>(deserializer: D) -> Result<KeypairOriginal, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let b = Vec::deserialize(deserializer)?;
+        KeypairOriginal::from_bytes(&b).map_err(serde::de::Error::custom)
+    }
+}
+
 #[pyclass(module = "solders.keypair", subclass)]
-#[derive(PartialEq, Debug)]
+#[derive(PartialEq, Debug, Serialize, Deserialize)]
 /// A vanilla Ed25519 key pair.
 ///
 /// Calling ``Keypair()`` creates a new, random ``Keypair``.
@@ -23,11 +44,11 @@ use crate::{
 ///     >>> from solders.keypair import Keypair
 ///     >>> assert Keypair() != Keypair()
 ///
-pub struct Keypair(pub KeypairOriginal);
+pub struct Keypair(#[serde(with = "keypair_serde")] pub KeypairOriginal);
 
 #[pyhash]
 #[richcmp_signer]
-#[common_magic_methods]
+#[common_methods]
 #[pymethods]
 impl Keypair {
     #[classattr]
@@ -214,7 +235,7 @@ impl PyFromBytesGeneral for Keypair {
     }
 }
 
-impl CommonMethods for Keypair {
+impl CommonMethods<'_> for Keypair {
     fn pystr(&self) -> String {
         self.0.to_base58_string()
     }