+++ src/transaction.rs
@@ -10,7 +10,7 @@ use solana_sdk::{
         TransactionError as TransactionErrorOriginal,
     },
 };
-use solders_macros::{common_magic_methods, richcmp_eq_only};
+use solders_macros::{common_methods, richcmp_eq_only};
 
 use crate::{
     convert_instructions, convert_optional_pubkey, handle_py_err, impl_display,
@@ -91,7 +91,7 @@ impl From<SanitizeErrorOriginal> for PyErrWrapper {
 pub struct Transaction(TransactionOriginal);
 
 #[richcmp_eq_only]
-#[common_magic_methods]
+#[common_methods]
 #[pymethods]
 impl Transaction {
     #[new]
@@ -568,7 +568,7 @@ impl RichcmpEqualityOnly for Transaction {}
 pybytes_general_via_bincode!(Transaction);
 py_from_bytes_general_via_bincode!(Transaction);
 impl_display!(Transaction);
-impl CommonMethods for Transaction {}
+impl CommonMethods<'_> for Transaction {}
 
 impl From<TransactionOriginal> for Transaction {
     fn from(tx: TransactionOriginal) -> Self {