+++ src/transaction_status.rs
@@ -1,0 +1,44 @@
+use pyo3::prelude::*;
+use serde::{Deserialize, Serialize};
+
+/// Encoding options for transaction data.
+#[pyclass(module = "solders.transaction_status")]
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub enum UiTransactionEncoding {
+    Binary, // Legacy. Retained for RPC backwards compatibility
+    Base64,
+    Base58,
+    Json,
+    JsonParsed,
+}
+
+impl Default for UiTransactionEncoding {
+    fn default() -> Self {
+        Self::Base64
+    }
+}
+
+/// Levels of transaction detail to return in RPC requests.
+#[derive(Debug, Clone, Copy, Eq, Hash, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+#[pyclass]
+pub enum TransactionDetails {
+    Full,
+    Signatures,
+    #[serde(rename = "none")]
+    None_,
+}
+
+impl Default for TransactionDetails {
+    fn default() -> Self {
+        Self::Full
+    }
+}
+
+pub fn create_transaction_status_mod(py: Python<'_>) -> PyResult<&PyModule> {
+    let m = PyModule::new(py, "transaction_status")?;
+    m.add_class::<TransactionDetails>()?;
+    m.add_class::<UiTransactionEncoding>()?;
+    Ok(m)
+}